Renamed ParseUtils to ExcelFile and added more helper functions
Renamed ParseUtils to ExcelFile and added more helper functions. Changed how the functions work to reating an object ExcelFile with a constructor that parses all the data in the Excel file to store a Workbook. Modified functions to get data from Workbook instead of opening and closing the Excel file each time to obtain information.