Created UseCase and Presenter for FetchUserData
Created a FetchUserDataUseCase to take in a SQL query SELECT statement or a List of constraints, where the first element is the table  name (doesn't work with other constraints). Gets the data and then returns it to the Presenter's ResultInterface. Created FetchUserDataPresenterImpl to implement FetchUserDataPresenter and FetchUserDataResultInterface to communicate between model and view. Create UserDataViews, but not implemted.