Completed ViewUserData User Interface
Completed the View Data part of the application. Can now select a template from a drop down of the existing templates and view the user data, requires there to be data in the database. Currently unable to show the table without clicking it to update. Deleted or modified some print statements to make the more informative.